+++ faq-bundle/src/Resources/contao/dca/tl_faq.php
@@ -422,7 +422,7 @@ class tl_faq extends Backend
 
 		return '
 <div class="cte_type ' . $key . '"><strong>' . $arrRow['question'] . '</strong> - ' . $date . '</div>
-<div class="limit_height' . (!Config::get('doNotCollapse') ? ' h52' : '') . '">
+<div class="limit_height' . (!Config::get('doNotCollapse') ? ' h40' : '') . '">
 ' . StringUtil::insertTagToSrc($arrRow['answer']) . '
 </div>' . "\n";
 	}