+++ faq-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Optimize the element preview height (see contao/core-bundle#678).
  * Use the file meta data by default when adding an image (see contao/core-bundle#807).
  * Add the DCA picker (see contao/core-bundle#755).
  * Improve the findByIdOrAlias() method (see contao/core-bundle#729).